Extract download error handling into app_error module
