+++ src/main.rs
@@ -1,11 +1,10 @@
+use crate::app_error::{api_error, to_local, AppError};
 use crate::parse_args::{parse_args, Commands, DownloadArgs, DownloadMode};
-use chrono::{
-    DateTime, Duration, Local, LocalResult, NaiveDate, NaiveDateTime, TimeZone, Timelike,
-};
-use std::fmt;
+use chrono::{DateTime, Duration, Local, NaiveDate, NaiveDateTime, Timelike};
 use std::path::{Path, PathBuf};
 use unifi_protect::*;
 
+mod app_error;
 mod parse_args;
 
 #[tokio::main]
@@ -22,93 +21,11 @@ async fn main() {
     }
 }
 
-#[derive(Debug)]
-enum AppError {
-    ParseDate {
-        input: String,
-        source: chrono::ParseError,
-    },
-    InvalidDateRange {
-        start: NaiveDateTime,
-        end: NaiveDateTime,
-    },
-    DateConstruction {
-        context: String,
-    },
-    DateConversion {
-        context: String,
-    },
-    DateOverflow {
-        context: String,
-    },
-    InvalidMode {
-        mode: String,
-    },
-    Api {
-        context: String,
-        source: String,
-    },
-}
-
-impl fmt::Display for AppError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            AppError::ParseDate { input, source } => {
-                write!(f, "failed to parse date '{}': {}", input, source)
-            }
-            AppError::InvalidDateRange { start, end } => {
-                write!(
-                    f,
-                    "invalid date range: end date/time '{}' is before start date/time '{}'",
-                    end, start
-                )
-            }
-            AppError::DateConstruction { context } => write!(f, "invalid date/time: {}", context),
-            AppError::DateConversion { context } => {
-                write!(f, "time zone conversion failed: {}", context)
-            }
-            AppError::DateOverflow { context } => write!(f, "date arithmetic failed: {}", context),
-            AppError::InvalidMode { mode } => write!(f, "invalid download mode '{}'", mode),
-            AppError::Api { context, source } => write!(f, "{}: {}", context, source),
-        }
-    }
-}
-
-impl std::error::Error for AppError {}
-
-fn to_local(naive: NaiveDateTime, context: String) -> Result<DateTime<Local>, AppError> {
-    match Local.from_local_datetime(&naive) {
-        LocalResult::Single(dt) => Ok(dt),
-        LocalResult::Ambiguous(first, second) => Err(AppError::DateConversion {
-            context: format!(
-                "{} ({}) is ambiguous between '{}' and '{}'",
-                context, naive, first, second
-            ),
-        }),
-        LocalResult::None => Err(AppError::DateConversion {
-            context: format!("{} ({}) does not exist in local timezone", context, naive),
-        }),
-    }
-}
-
-fn api_error(context: impl Into<String>, source: impl Into<String>) -> AppError {
-    AppError::Api {
-        context: context.into(),
-        source: source.into(),
-    }
-}
-
 async fn download(args: &DownloadArgs) -> Result<(), AppError> {
-    let start_date =
-        parse_date_or_hour(&args.start_date, true).map_err(|source| AppError::ParseDate {
-            input: args.start_date.clone(),
-            source,
-        })?;
-    let end_date =
-        parse_date_or_hour(&args.end_date, false).map_err(|source| AppError::ParseDate {
-            input: args.end_date.clone(),
-            source,
-        })?;
+    let start_date = parse_date_or_hour(&args.start_date, true)
+        .map_err(|source| AppError::parse_date(&args.start_date, source))?;
+    let end_date = parse_date_or_hour(&args.end_date, false)
+        .map_err(|source| AppError::parse_date(&args.end_date, source))?;
 
     if end_date < start_date {
         return Err(AppError::InvalidDateRange {
